Restructure parsers into package
