+++ parsers/__init__.py
@@ -1,0 +1,13 @@
+from .base import BaseBankParser
+from .generic import GenericEnglishParser
+from .factory import BankParserFactory
+from .spain.base import SpanishBankParser
+from .spain.santander import SantanderParser
+from .spain.bbva import BBVAParser
+from .spain.caixabank import CaixaBankParser
+from .argentina.galicia import GaliciaParser
+
+__all__ = [
+    'BaseBankParser', 'SpanishBankParser', 'SantanderParser', 'BBVAParser',
+    'CaixaBankParser', 'GaliciaParser', 'GenericEnglishParser', 'BankParserFactory'
+]
