+++ parsers/argentina/__init__.py
@@ -1,0 +1,3 @@
+from .galicia import GaliciaParser
+
+__all__ = ['GaliciaParser']
