+++ parsers/spain/__init__.py
@@ -1,0 +1,6 @@
+from .base import SpanishBankParser
+from .santander import SantanderParser
+from .bbva import BBVAParser
+from .caixabank import CaixaBankParser
+
+__all__ = ['SpanishBankParser', 'SantanderParser', 'BBVAParser', 'CaixaBankParser']
