+++ parsers/spain/bbva.py
@@ -1,0 +1,10 @@
+from .base import SpanishBankParser
+
+class BBVAParser(SpanishBankParser):
+    """Parser de extractos para BBVA."""
+
+    bank_id = 'bbva'
+    aliases = ['bbva']
+
+    def _get_bank_name(self) -> str:
+        return "BBVA"
